+++ src/main/java/nesemulator/cpu/addressmode/Abx.java
@@ -12,7 +12,7 @@ public class Abx extends AddressingMode {
     }
 
     @Override
-    public short set() {
+    public byte set() {
         return 0;
     }
 }
