+++ src/main/java/nesemulator/cpu/addressmode/Ind.java
@@ -12,7 +12,7 @@ public class Ind extends AddressingMode {
     }
 
     @Override
-    public short set() {
+    public byte set() {
         return 0;
     }
 }
