+++ src/main/java/nesemulator/cpu/addressmode/Izx.java
@@ -12,7 +12,7 @@ public class Izx extends AddressingMode {
     }
 
     @Override
-    public short set() {
+    public byte set() {
         return 0;
     }
 }
