+++ src/main/java/nesemulator/cpu/addressmode/Rel.java
@@ -12,7 +12,7 @@ public class Rel extends AddressingMode {
     }
 
     @Override
-    public short set() {
+    public byte set() {
         return 0;
     }
 }
