+++ src/main/java/nesemulator/cpu/opcodes/Bcs.java
@@ -5,7 +5,7 @@ package nesemulator.cpu.opcodes;
  */
 public class Bcs extends Opcode {
     @Override
-    public short operate() {
+    public byte operate() {
         return 0;
     }
 }
