+++ src/main/java/nesemulator/cpu/opcodes/Bvc.java
@@ -5,7 +5,7 @@ package nesemulator.cpu.opcodes;
  */
 public class Bvc extends Opcode {
     @Override
-    public short operate() {
+    public byte operate() {
         return 0;
     }
 }
